ProjetoWeb/requirements.txt
botcity-framework-web>=0.8.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0
requests
pandas
webdriver-manager==4.0.0
